Do not run make install clean script if there are uncommited changes

+++ tools/make_install_clean.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 # !Warning! Do not run when you have uncommmited changes.
-# TODO: Check if there is an uncommited changes.
+# https://github.com/robbyrussell/oh-my-zsh/blob/e55c715508a2f652fed741f2047c66dda2c6e5b0/lib/git.zsh
 # Compile, install the tinyos tools and clean the files.
-./Bootstrap && ./configure && make && sudo make install && git clean -df && git checkout -- .
+
+if [[ $(git status -s) == "" ]]
+then
+  ./Bootstrap && ./configure && make && sudo make install && git clean -df && git checkout -- .
+else
+  echo "There are uncommited changes."
+fi
